Determine :DEPLOY-CONSOLE feature at run time, not READ time
Previously, the :DEPLOY-CONSOLE feature had to be present before the system was
loaded in order for it to have any effect. In addition to forcing systems that
depend on deploy to set this flag as early as possible, it also means that if
:DEPLOY-CONSOLE is toggled after deploy is already in ASDF's .fasl cache its
effects won't be seen unless it is force loaded or removed the cache and
recompiled.

Now, its presence is checked at run time, working around the above issues.